Many changes + db in progress

+++ app/src/main/java/me/iscle/adifunofficial/station/entity/StationTrafficTypeCrossRef.kt
@@ -5,5 +5,5 @@ import androidx.room.Entity
 @Entity(primaryKeys = ["stationCode", "trafficTypeId"])
 data class StationTrafficTypeCrossRef(
     val stationCode: String,
-    val trafficTypeId: Int,
+    val trafficTypeId: Long,
 )
